+++ src/apps/Query/components/QuerySidesheet/QuerySidesheet.tsx
@@ -3,6 +3,7 @@ import { SideSheetContainer, SidesheetHeaderContent } from '@equinor/GardenUtils
 import { proCoSysUrls } from '@equinor/procosys-urls';
 import { SidesheetApi } from '@equinor/sidesheet';
 import { useEffect, useState } from 'react';
+import styled from 'styled-components';
 import { Query } from '../../types';
 import { DetailsTab, CommpkgTab } from './Tabs';
 
@@ -10,7 +11,9 @@ type QuerySideSheetProps = {
     item: Query;
     actions: SidesheetApi;
 };
-
+const SubHeader = styled.div`
+    margin: 0 8px 8px;
+`;
 export const QuerySideSheet = ({ item, actions }: QuerySideSheetProps): JSX.Element => {
     const [activeTab, setActiveTab] = useState<number>(0);
     const handleChange = (index: number) => {
@@ -19,11 +22,12 @@ export const QuerySideSheet = ({ item, actions }: QuerySideSheetProps): JSX.Elem
     const queryProcosysUrl = proCoSysUrls.getQueryUrl(item.queryUrlId ?? '');
 
     useEffect(() => {
-        actions.setTitle(<SidesheetHeaderContent title={item.queryUrlId} url={queryProcosysUrl} />);
-    }, [item.queryUrlId, queryProcosysUrl]);
+        actions.setTitle(<SidesheetHeaderContent title={item.queryNo} url={queryProcosysUrl} />);
+    }, [item.queryNo, queryProcosysUrl]);
 
     return (
         <SideSheetContainer>
+            <SubHeader> {item.title}</SubHeader>
             <Tabs activeTab={activeTab} onChange={handleChange}>
                 <Tabs.List>
                     <Tabs.Tab>Details</Tabs.Tab>